get cash advance endpoint added
