+++ src/Application/Queries/GetPurchaseOrder.cs
@@ -1,6 +1,4 @@
 ﻿using Application.Repositories;
-using Domain.Entities.Aggregates.PurchaseOrderAggregate;
-using Domain.Entities.Common;
 using Domain.Enums;
 using Domain.Exceptions;
 using Mapster;
